Change int to ll for large number handling ; Easy 01

+++ Easy/Palindrome-Number.cpp
@@ -25,9 +25,9 @@ public:
     bool isPalindrome(int x) {
         if (x < 0) return false;
 
-        int rev = 0, n = x;
+        ll rev = 0, n = x;
         while (x > 0) {
-            int d = x % 10;
+            ll d = x % 10;
             rev = rev * 10 + d;
             x /= 10;
         }
